Moved comp form type
